requirements.txt
crewai>=0.30.0
crewai-tools>=0.4.0
python-dotenv>=1.0.1
streamlit>=1.35
pydantic>=2.0.0
